refactor: retrieve cgo compilation var by specifying tag at build time
