+++ cmd/client/cmd.go
@@ -1,16 +1,12 @@
 package main
 
 import (
-	"fmt"
-	"os"
-	"runtime"
-
-	"github.com/juicity/juicity/config"
 	"github.com/spf13/cobra"
 )
 
 var (
-	rootCmd = &cobra.Command{
+	cgoEnabled = 0
+	rootCmd    = &cobra.Command{
 		Use:   "juicity-client [flags] [command [argument ...]]",
 		Short: "juicity-client is a quic-based proxy client.",
 		Long:  "juicity-client is a quic-based proxy client.",
@@ -21,19 +17,7 @@ var (
 )
 
 func init() {
-	rootCmd.AddCommand(&cobra.Command{
-		Use:   "version",
-		Short: "Print out version info",
-		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Printf("juicity-client version %v\n", config.Version)
-			fmt.Printf("go version %v %v/%v\n", runtime.Version(), runtime.GOOS, runtime.GOARCH)
-			if val, isSet := os.LookupEnv("CGO_ENALBED"); !isSet {
-				fmt.Print("CGO_ENALBED: NOT SET\n")
-			} else {
-				fmt.Printf("CGO_ENALBED: %v\n", val)
-			}
-		},
-	})
+	rootCmd.AddCommand()
 }
 
 // Execute executes the root command.
