+++ cmd/client/run.go
@@ -11,6 +11,7 @@ import (
 	"net"
 	"os"
 	"os/signal"
+	"runtime"
 	"syscall"
 	"time"
 
@@ -32,6 +33,16 @@ var (
 	logger = log.NewLogger(&log.Options{
 		TimeFormat: time.DateTime,
 	})
+
+	versionCmd = &cobra.Command{
+		Use:   "version",
+		Short: "Print out version info",
+		Run: func(cmd *cobra.Command, args []string) {
+			fmt.Printf("juicity-client version %v\n", config.Version)
+			fmt.Printf("go version %v %v/%v\n", runtime.Version(), runtime.GOOS, runtime.GOARCH)
+			fmt.Printf("CGO_ENALBED: %v\n", cgoEnabled)
+		},
+	}
 
 	runCmd = &cobra.Command{
 		Use:   "run",
@@ -166,5 +177,7 @@ func Serve(conf *config.Config) error {
 func init() {
 	// cmds
 	rootCmd.AddCommand(runCmd)
+	rootCmd.AddCommand(versionCmd)
+
 	shared.InitArgumentsFlags(runCmd)
 }
