+++ cmd/server/cgo_check.go
@@ -1,0 +1,8 @@
+//go:build cgo
+// +build cgo
+
+package main
+
+func init() {
+	cgoEnabled = 1
+}
